CSTE-package {CSTE}	R Documentation

Covariate specific treatment effect (CSTE) curve.

Description

Estimation of optimal individualized treatment rules using a covariate specific treatment effect (CSTE) curve. Suppose we have a binary treatment Z, a p-dimensional covariate X, which may be high-dimensional.

If the outcome Y is a binary variable, we consider the CSTE curve, defined by

CSTE(x) = logit( E(Y(1)\mid X=x)) - logit(E(Y(0)\mid X=x)),

where logit(u) = \log(u) -\log(1-u), Y(1) and Y(0) denote the potential outcomes if the active treatment and the control treatment are received, respectively. In this case, the CSTE curve represents the difference of the logarithm of odds ratio between treated group and control group, which is a common causal quantity in clinical studies for binary outcome. Moreover, Y = Z Y(1) + (1-Z) Y(0). Under the unconfoundedness assumption such that (Y(0), Y(1)) \perp Z \mid X, the CSTE curve can be re-expressed as

CSTE(x) = logit(E(Y \mid X=x, Z=1)) - logit( E(Y \mid X=x, Z=1)).

Denoting μ(X, Z) = E(Y| X, Z), we model the logarithm of odds ration as

logit(μ(X, Z)) = g_1(Xβ_1)Z + g_2(Xβ_2),

Then we have

CSTE(x) = g_1(x β_1).

If the outcome T is a failure time with right censoring, we let T(z) be the failure time under the corresponding treatment arm Z=z (z=0,1), and define the potential conditional hazard rate function as

λ^{(z)}(t\mid x) = \lim_{Δ t\to 0}\frac{P(t<T(z)≤q t+Δ t\mid T(z)>t, X=x)}{Δ t}.

In most medical research with a survival outcome, the treatment effect is usually represented by the logarithm of a hazard ratio. Hence, the CSTE curve at a fixed time t is defined as

CSTE(x,t) = \log≤ft\{\frac{λ^{(1)}(t\mid x)}{λ^{(0)}(t\mid x)}\right\}.

We focus on a special case that CSTE(x,t) = CSTE(x) in a randomized clinical trial. This assumption holds under many well-known models, such as Cox’s model and varying-coefficient Cox’s model. Consequently, we assume the following varying coefficient proportional hazard regression model for event time

λ(t| X, Z) = λ_0(t) \exp(g_1(Xβ_1)Z + g_2(Xβ_2)),

which implies that

CSTE(x) = g_1(x β_1)

This package facilitate estimating, making inference and predicting CSTE(x) for binary outcome and time to event outcome with right censoring.

Details

The R package CSTE - version 1.0 can be used for three main tasks:

cste_bin: estimate the CSTE curve for binary outcome.

cste_surv: estimate the CSTE curve for time to event outcome with right censoring.

predict_cste_bin: predict the CSTE curve of new subject with binary outcome.

predict_cste_surv: predict the CSTE curve of new subject for time to event outcome with right censoring.

cste_bin_SCB: compute the simultaneous confidence bands (SCB) of CSTE curve for binary outcome.

cste_surv_PCI: compute the point confidence intervals (PCI) of CSTE curve for time to event outcome with right censoring.

Author(s)

Peng Wu, Wenjie Hu, Yuhao Deng, Xiaohua Zhou.

Maintainer: Wenjie Hu <huwenjie@pku.edu.cn>

References

Guo W., Zhou X. and Ma S. (2021). Estimation of Optimal Individualized Treatment Rules Using a Covariate-Specific Treatment Effect Curve with High-dimensional Covariates, Journal of the American Statistical Association, 116(533), 309-321.

Ma Y. and Zhou X. (2017). Treatment selection in a randomized clinical trial via covariate-specific treatment effect curves, Statistical Methods in Medical Research, 26(1), 124-141.

[Package CSTE version 1.0 Index]