cste_bin_SCB {CSTE}	R Documentation

Calculate simultaneous confidence bands of CSTE curve for binary outcome.

Description

This function calculates simultaneous confidence bands of CSTE curve for binary outcome.

Usage

cste_bin_SCB(x, fit, alpha = 0.05)

Arguments

x
samples of predictor, which is a m*p matrix.

fit
a S3 class of cste.

alpha
the simultaneous confidence bands are of 1-α confidence level.

Value

A list which includes:

or_x: the ordered value of Xβ_1.

fit_x: the fitted value of CSTE curve corresponding to or_x.

lower_bound: the lower bound of CSTE's simultaneous confidence band.

upper_bound: the upper bound of CSTE's simultaneous confidence band.

References

Guo W., Zhou X. and Ma S. (2021). Estimation of Optimal Individualized Treatment Rules Using a Covariate-Specific Treatment Effect Curve with High-dimensional Covariates, Journal of the American Statistical Association, 116(533), 309-321

See Also

cste_bin

[Package CSTE version 1.0 Index]