cste_surv {CSTE}	R Documentation

Estimate the CSTE curve for time to event outcome with right censoring.

Description

Estimate the CSTE curve for time to event outcome with right censoring. The working model is

λ(t| X, Z) = λ_0(t) \exp(g_1(Xβ_1)Z + g_2(Xβ_2)),

which implies that CSTE(x) = g_1(xβ_1).

Usage

cste_surv(
  x,
  y,
  z,
  status = NULL,
  beta_ini = NULL,
  lam = 0,
  nknots = 2,
  max.iter = 200,
  eps = 0.001
)

Arguments

x
samples of covariates which is a n*p matrix.

y
samples of time to event which is a n*1 vector.

z
samples of treatment indicator which is a n*1 vector.

status
samples of censoring indicator which is a n*1 vector, default value is NULL, indicating no censoring.

beta_ini
initial values for (β_1', β_2')', default value is NULL.

lam
value of the lasso penalty parameter λ for β_1 and β_2, default value is 0.

nknots
number of knots for the B-spline for estimating g_1 and g_2.

max.iter
maximum iteration for the algorithm.

eps
numeric scalar ≥q 0, the tolerance for the estimation of β_1 and β_2.

Value

A S3 class of cste, which includes:

beta1: estimate of β_1.

beta2: estimate of β_2.

B1: the B-spline basis for estimating g_1.

B2: the B-spline basis for estimating g_2.

delta1: the coefficient of B-spline for estimating g_1.

delta2: the coefficient for B-spline for estimating g_2.

iter: number of iteration.

g1: the estimate for g_1(Xβ_1).

g2: the estimate for g_2(Xβ_2).

References

Ma Y. and Zhou X. (2017). Treatment selection in a randomized clinical trial via covariate-specific treatment effect curves, Statistical Methods in Medical Research, 26(1), 124-141.

See Also

cste_surv_PCI, predict_cste_surv, select_cste_surv

Examples

## Quick example for the cste

# --------  Example 1: Simulated Data ---------  # 
## generate data 
set.seed(100)
X1 <- runif(200,0,20)
X2 <- runif(200,0,20)
X3 <- runif(200,0,20)
X <- cbind(X1, X2, X3)
Z <- c(rep(1,100),rep(0,100))

beta <- c(0.3,0.4,0.5)
lambda <- 0.1 * exp(log(X%*%beta/10+0.6)*Z+(X%*%beta)/20)
Time <- -log(runif(200,0,1))/lambda
C <- runif(200,12,15)
S <- as.numeric(Time <= C)
Time <- Time*S + C*(1-S)

## estimate the CSTE curve
fit <- cste_surv(X, Time, Z, S, nknots=3, max.iter=200)

## pointwise confidence interval (PCI)
res <- cste_surv_PCI(fit)

## plot
plot(res$or_x, res$fit_x, col = 'red', type = "l", 
     xlim=c(8,18), ylim=c(-1,3), lwd = 2, 
     ylab = "CSTE", xlab = "X * beta1",
     main ="Pointwise Confidence interval")
lines(res$or_x, res$lower_bound, lwd = 3, col = 'purple', lty = 2)
lines(res$or_x, res$upper_bound, lwd = 3, col = 'purple', lty = 2)
abline(h = 0, lty = 2, cex = 0.2)
legend("topleft", legend = c("Estimates", "PCI"), 
        lwd = c(2,3), lty = c(1, 2), col = c('red','purple'))
# xb <- sort(as.numeric(X%*%beta))
# lines(xb, log(xb/10+0.6), type='l', col='blue')

# --------  Example 2: Real Data ---------  # 
dat <- survival::lung

## estimate the CSTE curve
fit <- cste_surv(dat[,6:8], dat$time, dat$sex, 
                 dat$status, nknots = 3, max.iter = 100)
                 
## 95% pointwise confidence interval (PCI)
res <- cste_surv_PCI(fit)

## plot
plot(res$or_x, res$fit_x, col = 'red', type = "l", lwd = 2, ylim = c(-5,2),
     ylab = "CSTE", xlab = "X * beta1", main ="Pointwise Confidence interval")
lines(res$or_x, res$lower_bound, lwd = 3, col = 'purple', lty = 2)
lines(res$or_x, res$upper_bound, lwd = 3, col = 'purple', lty = 2)
abline(h = 0, lty = 2, cex = 0.2)
legend("topleft", legend = c("Estimates", "PCI"), 
        lwd = c(2,3), lty = c(1, 2), col = c('red','purple'))

[Package CSTE version 1.0 Index]