cste_surv_PCI {CSTE}	R Documentation

Calculate pointwise confidence intervals of CSTE curve for time to event outcome with right censoring.

Description

This function calculates pointwise confidence intervals of CSTE curve for time to event outcome with right censoring.

Usage

cste_surv_PCI(fit, alpha = 0.05)

Arguments

fit
a S3 class of cste.

alpha
the pointwise confidence intervals are of 1-α confidence level.

Value

A list which includes:

or_x: the ordered value of Xβ_1.

fit_x: the fitted value of CSTE curve corresponding to or_x.

lower_bound: the lower bound of CSTE's pointwise confidence intervals.

upper_bound: the upper bound of CSTE's pointwise confidence intervals.

References

Ma Y. and Zhou X. (2017). Treatment selection in a randomized clinical trial via covariate-specific treatment effect curves, Statistical Methods in Medical Research, 26(1), 124-141.

See Also

cste_surv

[Package CSTE version 1.0 Index]