predict_cste_bin {CSTE}	R Documentation

Predict the CSTE curve of new data for binary outcome.

Description

Predict the CSTE curve of new data for binary outcome.

Usage

predict_cste_bin(obj, newx)

Arguments

obj
a S3 class of cste.

newx
samples of covariates which is a m*p matrix.

Value

A S3 class of cste which includes

g1: predicted g_1(Xβ_1).

g2: predicted g_2(Xβ_2).

B1: the B-spline basis for estimating g_1.

B2: the B-spline basis for estimating g_2.

References

Guo W., Zhou X. and Ma S. (2021). Estimation of Optimal Individualized Treatment Rules Using a Covariate-Specific Treatment Effect Curve with High-dimensional Covariates, Journal of the American Statistical Association, 116(533), 309-321

See Also

cste_bin

[Package CSTE version 1.0 Index]