predict_cste_surv {CSTE}	R Documentation

Predict the CSTE curve of new data for time to event outcome with right censoring.

Description

Predict the CSTE curve of new data for time to event outcome with right censoring.

Usage

predict_cste_surv(obj, newx, alpha = 0.05)

Arguments

obj
a S3 class of cste.

newx
samples of covariates which is a m*p matrix.

alpha
(1-α) confidence level.

Value

A S3 class of cste, which includes

g1: predicted g_1(Xβ_1).

lower_bound: the lower bound of CSTE's pointwise confidence intervals.

upper_bound: the upper bound of CSTE's pointwise confidence intervals.

References

Ma Y. and Zhou X. (2017). Treatment selection in a randomized clinical trial via covariate-specific treatment effect curves, Statistical Methods in Medical Research, 26(1), 124-141.

See Also

cste_surv

[Package CSTE version 1.0 Index]