select_cste_surv {CSTE}	R Documentation

Select the optimal tuning parameters in CSTE estimation for time to event outcome with right censoring.

Description

select lasso penalty parameter λ for β_1 and β_2 in CSTE estimation.

Usage

select_cste_surv(
  x,
  y,
  z,
  status = NULL,
  lam_seq,
  beta_ini = NULL,
  nknots = 1,
  max.iter = 2000,
  eps = 0.001
)

Arguments

x
samples of covariates which is a n*p matrix.

y
samples of binary outcome which is a n*1 vector.

z
samples of treatment indicator which is a n*1 vector.

status
samples of censoring indicator which is a n*1 vector. Default value is NULL, indicating no censoring.

lam_seq
a sequence for the choice of λ.

beta_ini
initial values for (β_1', β_2')', default value is NULL.

nknots
number of knots for the B-spline for estimating g_1 and g_2.

max.iter
maximum iteration for the algorithm.

eps
numeric scalar ≥q 0, the tolerance for the estimation of β_1 and β_2.

Value

A list which includes

optimal: optimal cste within the given the sequence of λ.

bic: BIC for the sequence of λ.

lam_seq: the sequence of λ that is used.

References

Ma Y. and Zhou X. (2017). Treatment selection in a randomized clinical trial via covariate-specific treatment effect curves, Statistical Methods in Medical Research, 26(1), 124-141.

See Also

cste_surv

[Package CSTE version 1.0 Index]